backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Flask-CORS
PyJWT
Werkzeug
python-dotenv
mysqlclient
scikit-learn
pandas
numpy 